requirements.txt
psycopg2
pandas
plotly
dash
python-dotenv
